# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PanelDomainResponse": {
        "properties": {
          "cookie_secure": {
            "title": "Cookie Secure",
            "type": "boolean"
          },
          "domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "domain",
          "cookie_secure",
          "url"
        ],
        "title": "PanelDomainResponse",
        "type": "object"
      },
...
      "ServiceStatusResponse": {
        "properties": {
          "active_state": {
            "title": "Active State",
            "type": "string"
          },
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "enabled": {
            "title": "Enabled",
            "type": "string"
          },
          "sub_state": {
            "title": "Sub State",
            "type": "string"
          },
          "unit": {
            "title": "Unit",
            "type": "string"
          }
        },
        "required": [
          "unit",
          "available",
          "active_state",
          "sub_state",
          "enabled"
        ],
        "title": "ServiceStatusResponse",
        "type": "object"
      },
      "SetPanelDomainRequest": {
        "properties": {
          "domain": {
            "maxLength": 253,
            "minLength": 1,
            "title": "Domain",
            "type": "string"
          },
          "force": {
            "default": false,
            "title": "Force",
            "type": "boolean"
          }
        },
        "required": [
          "domain"
        ],
        "title": "SetPanelDomainRequest",
        "type": "object"
      },
...
      "SystemStatsResponse": {
        "properties": {
          "cpu_percent": {
            "title": "Cpu Percent",
            "type": "number"
          },
          "disk_percent": {
            "title": "Disk Percent",
            "type": "number"
          },
          "disk_total": {
            "title": "Disk Total",
            "type": "integer"
          },
          "disk_used": {
            "title": "Disk Used",
            "type": "integer"
          },
          "load_avg": {
            "items": {
              "type": "number"
            },
            "title": "Load Avg",
            "type": "array"
          },
          "memory_percent": {
            "title": "Memory Percent",
            "type": "number"
          },
          "memory_total": {
            "title": "Memory Total",
            "type": "integer"
          },
          "memory_used": {
            "title": "Memory Used",
            "type": "integer"
          },
          "uptime_seconds": {
            "title": "Uptime Seconds",
            "type": "integer"
          }
        },
        "required": [
          "cpu_percent",
          "load_avg",
          "memory_total",
          "memory_used",
          "memory_percent",
          "disk_total",
          "disk_used",
          "disk_percent",
          "uptime_seconds"
        ],
        "title": "SystemStatsResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "HostyPanel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/system/panel-domain": {
      "delete": {
        "description": "Back to IP-only access: remove the vhost and allow cookies over HTTP.",
        "operationId": "clear_panel_domain_api_system_panel_domain_delete",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PanelDomainResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Clear Panel Domain",
        "tags": [
          "system"
        ]
      },
      "get": {
        "operationId": "get_panel_domain_api_system_panel_domain_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PanelDomainResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Panel Domain",
        "tags": [
          "system"
        ]
      },
      "put": {
        "description": "Serve the panel on a domain over HTTPS.\n\nPublishes the vhost to Caddy (which obtains the certificate), persists\nHOSTY_PANEL_DOMAIN + HOSTY_COOKIE_SECURE to the env file, and enables\nsecure cookies immediately. The current plain-HTTP session keeps working\nuntil its access token expires; log in again on the HTTPS URL.",
        "operationId": "set_panel_domain_api_system_panel_domain_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPanelDomainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PanelDomainResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Panel Domain",
        "tags": [
          "system"
        ]
      }
    },
    "/api/system/services": {
      "get": {
        "operationId": "list_services_api_system_services_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ServiceStatusResponse"
                  },
                  "title": "Response List Services Api System Services Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Services",
        "tags": [
          "system"
        ]
      }
    },
    "/api/system/services/{unit}": {
      "get": {
        "operationId": "service_status_api_system_services__unit__get",
        "parameters": [
          {
            "in": "path",
            "name": "unit",
            "required": true,
            "schema": {
              "title": "Unit",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Service Status",
        "tags": [
          "system"
        ]
      }
    },
    "/api/system/services/{unit}/actions/{action}": {
      "post": {
        "operationId": "service_action_api_system_services__unit__actions__action__post",
        "parameters": [
          {
            "in": "path",
            "name": "unit",
            "required": true,
            "schema": {
              "title": "Unit",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "action",
            "required": true,
            "schema": {
              "title": "Action",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Service Action",
        "tags": [
          "system"
        ]
      }
    },
    "/api/system/stats": {
      "get": {
        "operationId": "system_stats_api_system_stats_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemStatsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "System Stats",
        "tags": [
          "system"
        ]
      }
    },
...
  }
}